feat(lang): i64 through the pipeline — literals, i64(x), arith/cmp
Slice 2 of the language-capability track (prereq for Hash<K,i64> state
and time.nowMicros). Surface, as user-approved:

- Int literals accumulate to i64 in the parser (cap INT64_MAX positive,
  2^63 fused-negative magnitude); analyze types them I32 when they fit
  i32, else I64 — `60000000000` just works, beyond i64 is InvalidInteger.
- `i64(x)` conversion builtin (Swift-initializer style): folds literals,
  identity on i64, WidenI64→SextI64→LLVMBuildSExt on runtime i32;
  shadowed by a user function named i64; works as a pipe stage (x | i64(_)).
- Arithmetic and comparisons are same-width over {i32,i64}; mixed
  non-literal widths error with an i64(x) fix-it; a bare int literal
  adopts the i64 side (so `i64(x) + 1` and `-i64(x)` work — the parser's
  unary-minus desugar Sub(0,x) depends on this).
- Same total semantics at 64 bits: wrap, x/0==0, INT64_MIN/-1==INT64_MIN,
  INT64_MIN%-1==0. fold_arith_i64 mirrors fold_arith_i32; codegen div/mod
  guards are width-aware (constants follow the operand's RIR type).
- const_eval handles I64 literals/arith/compares and WidenI64, so
  `if const` folds 64-bit predicates.

Ripple from widening int_value (AST/HIR/MIR/ConstValue i32→i64):
u32-bound guards added for WS frame-len bounds and wait(timer(ms));
seen-int-case dedup widened; variant/error tag narrowings made explicit.

Deliberately out of scope, each locked by a rejection test: match on an
i64 subject (all six subject-analysis sites gated), bitwise.* over i64,
i64 type annotations (unnameable, like ByteSize), narrowing conversion.
Yield slots are NOT needed: locals re-materialize in resume states (slots
only carry wait metadata) — pinned by the i64-across-wait JIT test.

Tests: 12 frontend + 3 JIT (threshold matrix incl. INT64 bounds, builtin
fold/shadowing/pipe, boundary folds, literal adoption, mixed-width and
match/bitwise/annotation rejections, if-const i64, runtime arith/compare/
div-mod guards crossing 2^31 and 2^63, i64 local across wait).

Docs: DESIGN.md §3.2.1 integer-widths block, language-card i64(x) idiom,
CLAUDE.md bullets.

+++ include/rut/compiler/ast.h
@@ -175,7 +175,7 @@ struct AstExpr {
     AstExprKind kind = AstExprKind::BoolLit;
     Span span{};
     bool bool_value = false;
-    i32 int_value = 0;
+    i64 int_value = 0;
     Str str_value{};
     Str msg{};
     Str name{};
